checkbox students

+++ WLANCodesWebApp/views.py
@@ -130,41 +130,44 @@ def students(request):
             students = Student.objects.all().order_by('group', 'name')
             return render(request, 'students.html', {'students': students})
     if request.method == 'POST':
-        id = int(request.POST.get('send'))
-        student = Student.objects.get(id=id)
-        oldcode = student.code
-        # put oldcode on delete list
-        if oldcode is not None:
-            CodeDeletion.objects.create(
-                code_to_delete=oldcode,
-                name=student.name,
-                firstname=student.firstname,
-                group=student.group
+        if request.POST.get('checked'):
+            print(request.POST.getlist('checkbox'))
+        else:
+            id = int(request.POST.get('send'))
+            student = Student.objects.get(id=id)
+            oldcode = student.code
+            # put oldcode on delete list
+            if oldcode is not None:
+                CodeDeletion.objects.create(
+                    code_to_delete=oldcode,
+                    name=student.name,
+                    firstname=student.firstname,
+                    group=student.group
+                )
+            newcode = Code.objects.filter(type='y').first()
+            student.code = newcode.code
+            # delete used code
+            newcode.delete()
+            student.date = datetime.today()
+            student.save()
+            mail_text = (
+                "Hallo " + student.firstname + ",\n\n" +
+                "hiemit erhältst du deinen WLAN-Code für das aktuelle Schuljahr. " +
+                "Der Code kann nur einmalig auf einem Gerät aktiviert werden, d.h. " +
+                "falls du ein Tablet hast, dein Tablet, ansonsten dein Smartphone.\n\n" +
+                "Dein Code lautet: \n\n" +
+                student.code + "\n\n" +
+                "Hinweis: Eine Weitergabe des Codes ist nicht möglich. Falls du einen " +
+                "neuen Code brauchst, wird der alte Code deaktiviert."
             )
-        newcode = Code.objects.filter(type='y').first()
-        student.code = newcode.code
-        # delete used code
-        newcode.delete()
-        student.date = datetime.today()
-        student.save()
-        mail_text = (
-            "Hallo " + student.firstname + ",\n\n" +
-            "hiemit erhältst du deinen WLAN-Code für das aktuelle Schuljahr. " +
-            "Der Code kann nur einmalig auf einem Gerät aktiviert werden, d.h. " +
-            "falls du ein Tablet hast, dein Tablet, ansonsten dein Smartphone.\n\n" +
-            "Dein Code lautet: \n\n" +
-            student.code + "\n\n" +
-            "Hinweis: Eine Weitergabe des Codes ist nicht möglich. Falls du einen " +
-            "neuen Code brauchst, wird der alte Code deaktiviert."
-        )
-        noreply = Config.objects.get(name="noreply-mail")
-        send_mail(
-            'WLAN-CODE',
-            mail_text,
-            noreply,
-            [student.email],
-            fail_silently=True,
-        )
+            noreply = Config.objects.get(name="noreply-mail")
+            send_mail(
+                'WLAN-CODE',
+                mail_text,
+                noreply,
+                [student.email],
+                fail_silently=True,
+            )
         return redirect('students')
